+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 0.2.9 (????-??-??)
 ------------------
 
+- method `JavaObject.new_instance` in module `weka.core.classes` now automatically
+  installs packages based on suggestions if the JVM was started with the `auto_install`
+  flag enabled.
 - ...